Allow building on case-insensitive filesystems
Previously, we had both exif.c (part of of the Eric Johnston's exiftags
library) and EXIF.xs, from which EXIF.c is generated.  That's fine on
case-sensitive filesystems, but on typical Win32 and Mac OS systems (and
VMS, presumably), the two .c filenames conflict.

Fix by using Image-EXIF.xs as the name of the XS file; that means we
generate Image-EXIF.c from it, which won't conflict.

This closes RT#4353, RT#4973, and RT#76281.