Delete XS-side "error" buffer
It never gets set to anything.

Since this means that c_errstr() could never return a true value, also
delete both c_errstr() and the code that attempted to handle a true return
from c_errstr().  That in turn allows us to replace the ->error and ->errstr
methods with trivial "everything is fine" implementations.